refactor(ngcc): move locking code into its own folder
